+++ src/common/collections/tenants/server/publish.js
@@ -52,7 +52,7 @@ Meteor.publish( TenantsManager.C.pub.tenantsAll.publish, async function(){
                         // on HMR, happens that Error: Could not find element with id wx8rdvSdJfP6fCDTy to change
                         self.added( TenantsManager.C.pub.tenantsAll.collection, entity._id, await Tenants.s.transformEntity( entity ));
                         TenantsManager.s.eventEmitter.emit( 'added', entity._id, await Tenants.s.transformEntity( entity ));
-                        console.debug( e, 'ignored' );
+                        //console.debug( e, 'ignored' );
                     }
                 } else {
                     console.warn( 'added: entity not found', item.entity );